testing prefetch

+++ src/vpux_compiler/include/vpux/compiler/core/runtime_simulator.hpp
@@ -61,7 +61,7 @@ public:
     };
 
     RuntimeSimulator(mlir::MLIRContext* ctx, mlir::FuncOp func, Logger log, int64_t numDmaEngines,
-                     size_t numRealBarriers);
+                     int64_t numRealBarriers);
     void init();
     bool assignPhysicalIDs();
     void buildTaskLists();
